Updating route checking to prevent unintended empty routes
Re: issue #26. Passing extra slash characters to the `rest_route` parameter was coming back as an empty `$current_route`, which meant we weren't raising an error when we should have been. It seems that the extra "if not empty" was not only unnecessary here, but actually causing this unintended exposure of the top-level endpoint of the REST API.