Updating for "do or don't manage routes" per role. Tagged 1.6-beta
This solved the issue we had when a role was "default allow" but nothing was defined yet. There was no difference between "undefined" and a user's desire to "not allow". Now we either "allow all" for a whole role OR manage routes, and what you see is what you get. Undefined ALWAYS automatically means blocked.